fusor server: logic to locate path to cloudforms ova file
This commit contains changes to programatically determine the
location of the CloudForms ova file, which has been synced to
a Katello/Pulp repository.

The general logic is:

- if user specifies the ova file in fusor.yaml, locate the path
  to that file (e.g. :image_file_name: "cfme-rhevm-5.3-54.x86_64.rhevm.ova")

- if the fusor.yaml does not include the :image_file_name, retrieve
  the path to the 'latest ova file (e.g. "cfme-rhevm-5.3-57.x86_64.rhevm.ova")

It is anticipated that the default behavior (using latest) will be
most desirable; however, the yaml file setting is supported,
in case a workaround is needed to enforce a particular version of
the file.